apps/api/pyproject.toml
[project]
name = "multi-agent-bench-api"
version = "0.1.0"
description = "FastAPI backend for the pharmacy multi-agent POC"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.0",
  "boto3>=1.34.0",
  "fastapi>=0.115.0",
  "pydantic>=2.7.0",
  "psycopg[binary]>=3.2.0",
  "python-dotenv>=1.0.0",
  "python-multipart>=0.0.9",
  "sse-starlette>=2.1.0",
  "sqlalchemy>=2.0.0",
  "uvicorn[standard]>=0.30.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.0.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
